+++ changelog.md
@@ -13,3 +13,8 @@
 * Move user rclone.conf & related to /sdcard/.rclone/
 * Control global --vfs-cache-mode via simple files placed in /sdcard/.rclone/
 * Specify custom params for individual remotes via /sdcard/.rclone/.REMOTENAME.params
+
+## v1.4
+* Add ability to disable a remote
+* Add a wrapper script for rclone
+* Make remount possible via `su --mount-master`